fix(platform): break the host-activity import cycle, classify it for the PAL coverage gate, and bump doc schema refs to V60
Addresses three preflight:fast gates that were red because of Task 1
(and, for #3, Task 2's schema bump):

1. audit:boundaries — host-activity.ts <-> host-activity/{linux,darwin,win32}.ts
   was a real runtime cycle (the backends import UNKNOWN_PROBE, a value,
   not just a type). Extracted the shared surface (HostPower,
   HostProbeSource, HostActivityProbe, HostActivity, UNKNOWN_PROBE,
   UNKNOWN_HOST_ACTIVITY) into a new leaf module,
   platform/host-activity-types.ts. The three backends now import from
   that leaf instead of host-activity.ts. host-activity.ts re-exports
   everything so existing importers (platform/types.ts,
   platform/assemble.ts, fleet/fleet-store.ts) are unaffected.

2. audit:coverage-gate-pal — the four new files branch on platform
   (three are OS-named, host-activity.ts imports `platform` from
   node:os) but were absent from PLATFORM_BRANCHING_ALLOWLIST. Added
   all four as gate: "none", matching the sibling entries for
   platform/index.ts, platform/{win32,darwin,linux}.ts: no
   coverage-threshold gate targets src/platform directly (the Sandbox
   gate's test path is scoped one directory down, to
   src/platform/sandbox), and no gate exercises the fleet scheduler
   yet either.

3. audit:status-drift — CLAUDE.md, GEMINI.md and docs/architecture.md
   each carried a bare "schema V59" ceiling line; canonical is now V60
   after Task 2's migration. Updated only that ceiling line in each
   file (the unrelated "schema **V59** `media_grant`" historical
   mention in CLAUDE.md/GEMINI.md is bold-emphasised prose describing
   a past PR and is untouched, both because it is historically
   accurate and because the audit's regex does not match it). Scope
   held to the version number only, per instruction: no I38/D28/fleet
   description added to these three files.
