+++ packages/gateway/src/platform/host-activity.ts
@@ -1,42 +1,19 @@
 import { platform } from "node:os";
+import type { HostActivity, HostActivityProbe } from "./host-activity-types.ts";
+import { UNKNOWN_PROBE } from "./host-activity-types.ts";
 
-/** Mains state. `unknown` is a real answer — a desktop or VM has no battery to report. */
-export type HostPower = "ac" | "battery" | "unknown";
-
-/**
- * How much of a probe is real. `power_only` means user-idle could not be measured on this host
- * (headless Linux, a CI runner with no HID) — the fleet still admits, and says so. It never
- * claims a check it did not perform.
- */
-export type HostProbeSource = "measured" | "power_only";
-
-export interface HostActivityProbe {
-  readonly power: HostPower;
-  /** Milliseconds since last user input, or null when genuinely unmeasurable. */
-  readonly idleMs: number | null;
-  readonly source: HostProbeSource;
-}
-
-export interface HostActivity {
-  probe(): Promise<HostActivityProbe>;
-}
-
-/** The probe returned when a backend cannot answer at all. Never throws upward. */
-export const UNKNOWN_PROBE: HostActivityProbe = Object.freeze({
-  power: "unknown",
-  idleMs: null,
-  source: "power_only",
-});
-
-/**
- * A shared stub for tests/fixtures that need a `HostActivity` value but do not exercise it.
- * `hostActivity` is a required `PlatformServices` member (never optional — an omitted field must
- * fail typecheck, not go silently inert), so every literal that predates this feature needs one of
- * these to keep compiling.
- */
-export const UNKNOWN_HOST_ACTIVITY: HostActivity = {
-  probe: async (): Promise<HostActivityProbe> => UNKNOWN_PROBE,
-};
+// Re-exported so existing importers (`platform/types.ts`, `platform/assemble.ts`,
+// `fleet/fleet-store.ts`) keep working unchanged — the types/values themselves live in the leaf
+// module `host-activity-types.ts` to avoid a runtime import cycle: the three backends below need
+// `UNKNOWN_PROBE` as a VALUE, and this file dynamically imports the backends, so a value defined
+// here and imported back by a backend would be a real cycle (`audit:boundaries` enforces this).
+export type {
+  HostActivity,
+  HostActivityProbe,
+  HostPower,
+  HostProbeSource,
+} from "./host-activity-types.ts";
+export { UNKNOWN_HOST_ACTIVITY, UNKNOWN_PROBE } from "./host-activity-types.ts";
 
 /**
  * Each platform answers for its own mechanism, so this knowledge stays in the PAL rather than
